Fix 2

+++ src/main/java/de/keks/internal/command/config/SetupConfig.java
@@ -112,9 +112,9 @@ public class SetupConfig {
 		ConfigValues.taxIncreasePerRegion = (double) setupPath(taxIncreasePerRegion, 10.0D);
 		ConfigValues.firstRegionCosts = (double) setupPath(firstRegionCosts, 256.0D);
 		ConfigValues.percentageAtRegionSell = (double) setupPath(percentageAtRegionSell, 0.5D);
-		ConfigValues.buyupMembers = (double) setupPath(buyupMembers, 20 * 24 * 60 * 60 * 1000);
-		ConfigValues.buyupNoMembers = (double) setupPath(buyupNoMembers, 35 * 24 * 60 * 60 * 1000);
-		ConfigValues.buyupInfo = (double) setupPath(buyupInfo, 10);
+		ConfigValues.buyupMembers = (int) setupPath(buyupMembers, 20 * 24 * 60 * 60 * 1000);
+		ConfigValues.buyupNoMembers = (int) setupPath(buyupNoMembers, 35 * 24 * 60 * 60 * 1000);
+		ConfigValues.buyupInfo = (double) setupPath(buyupInfo, 10D);
 		ConfigValues.landBuyChunkBorders = (Material) setupPath(landBuyChunkBorders, Material.TORCH);
 		ConfigValues.landSellChunkBorders = (Material) setupPath(landSellChunkBorders, Material.REDSTONE_TORCH_ON);
 		ConfigValues.setBiome = (double) setupPath(setBiome, 300D);
